1854-environment-exposure-calculation: refactor
