+++ Agents/EnvironmentExposureCalculationAgent/agent/route.py
@@ -1,13 +1,10 @@
-from datetime import date
-import json
-from flask import abort, jsonify, request
-import pandas as pd
+from flask import jsonify, request
 
-from agent.boundary import create_buffer_around_points
-from agent.exposure.intersect import Intersect
-from agent.output.output_as_json import output_as_json
+from agent.boundary import create_boundary
+from agent.exposure import exposure_calculation
+from agent.output import get_output
 from agent.output.save_to_table import save_to_table
-from agent.point_selection.point_selection import create_ponits_table_self_defined_area
+from agent.point_selection import select_points
 from agent.utils.table_name_helper import TableNameHelper
 
 
@@ -48,48 +45,3 @@ def register_route(app):
         
         return output
         
-
-def select_points(table_name:str, point_selection: str, data: dict) -> str:
-    if point_selection == 'selected_points':
-        points = data['points']  # list of points
-        # todo: need to fix the input
-        # create_points_table(points)
-    elif point_selection == 'region':
-        # get points in a preset region
-        region = data['region']
-        # todo: need to add this function
-        # create_points_table_region()
-    elif point_selection == 'self_defined_region':
-        lng_start = data.get('lng_start')
-        lng_end = data.get('lng_end')
-        lng_step = data.get('lng_step')
-        lat_start = data.get('lat_start')
-        lat_end = data.get('lat_end')
-        lat_step = data.get('lat_step')
-
-        create_ponits_table_self_defined_area(lng_start, lng_end, lng_step,
-                            lat_start, lat_end, lat_step, table_name)
-    else:
-        abort(400, description="Unsupported point selection method")
-    return table_name
-
-def create_boundary(points_table_name: str, boundary: str, data:dict):
-    if boundary == 'buffer':
-        boundary_radius = float(data['boundary_radius'])
-        create_buffer_around_points(points_table_name, boundary_radius)
-    else:
-        abort(400, description="Unsupported buffer method")
-
-def exposure_calculation(table_name_helper:TableNameHelper, algorithm: str) -> pd.DataFrame:
-    if algorithm == 'intersect':
-        return Intersect(table_name_helper).intersect()
-    else:
-        abort(400, description="Unsupported calculation algorithm method")
-
-def get_output(output_format: str, res:pd.DataFrame, table_name_helper:TableNameHelper):
-    if output_format == 'csv':
-        pass
-    elif output_format == 'json':
-        return output_as_json(res, table_name_helper)
-    else:
-        abort(400, description="Unsupported output format")
